+++ example/lib/src/modules/auth/auth_store.dart
@@ -1,9 +1,5 @@
 class AuthStore {
-  AuthStore() {
-    print('🔐 AuthStore criado');
-  }
+  AuthStore();
 
-  void dispose() {
-    print('🔐 AuthStore disposto');
-  }
+  void dispose() {}
 }