+++ example/lib/src/modules/home/home_module.dart
@@ -1,5 +1,3 @@
-import 'dart:async';
-import 'package:flutter/material.dart';
 import 'package:go_router_modular/go_router_modular.dart';
 import '../shared/shared_module.dart';
 import 'pages/home_page.dart';
@@ -7,9 +5,6 @@ import 'pages/demo_page.dart';
 
 class HomeModule extends Module {
   // Controle de estado do módulo
-  final bool _isInitialized = false;
-  Timer? _homeTimer;
-  StreamSubscription? _homeSubscription;
 
   @override
   List<Module> get imports {
@@ -38,14 +33,10 @@ class HomeModule extends Module {
   }
 
   @override
-  void initState(Injector i) {
-    print('init home module');
-  }
+  void initState(Injector i) {}
 
   @override
-  void dispose() {
-    print('dispose home module');
-  }
+  void dispose() {}
 }
 
 class HomeService {